dom/devicestorage/test/test_enumerateMultipleContinue.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html> <!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717103
-->
<head>
  <title>Test for the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717103">Mozilla Bug 717103</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

devicestorage_setup();

function enumerateSuccess(e) {
}

function enumerateFailure(e) {
}

var cursor = navigator.getDeviceStorage("testing")[0].enumerate();
cursor.onsuccess = enumerateSuccess;
cursor.onerror = enumerateFailure;

try {
 cursor.continue();
}
catch (e) {
  ok(true, "Calling continue before enumerateSuccess fires should throw");
  devicestorage_cleanup();
}


</script>
</pre>
</body>
</html>
